+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * It's now possible to schedule times to cycle through photos
 * It's now possible to restore recently deleted photos (photo widget only, for
   photos deleted after this update)
+* It's now possible to view the original photo when using the view in full
+  screen tap action
 
 ### Changed